requirements.txt
requests
sqlalchemy
fastapi
uvicorn
python-dotenv
pandas
apscheduler
pytz
rich
aiohttp
jinja2
fyers-apiv3
